Remove About and What's New dialogs
This commit removes the "About" dialog that was shown on first use,
and the "What's New" dialog that was shown on application updates.

It also removes the associated string resources and preference XML files
for licences and location settings that are no longer used.

Additionally, "ManualCamera" has been changed to "Manual Camera" in some
string resources for consistency.